Remove now-unnecessary cast
Original commit: elastic/x-pack-elasticsearch@fcbbe5a919fbb38ea47491e97c66e877d6578505

+++ plugin/src/main/java/org/elasticsearch/xpack/ml/datafeed/DatafeedConfig.java
@@ -248,7 +248,7 @@ public class DatafeedConfig extends AbstractDiffable<DatafeedConfig> implements 
         if (dateHistogram.dateHistogramInterval() != null) {
             return validateAndGetCalendarInterval(dateHistogram.dateHistogramInterval().toString());
         } else {
-            return (long) dateHistogram.interval();
+            return dateHistogram.interval();
         }
     }
